[federation] Move vals and functions to internal scope
To cleanup our public class, function, and value exposure, I moved some of the federation objects to internal scope since they do not need to be exposed as public
