+++ .gitignore
@@ -7,6 +7,7 @@ pom.xml.versionsBackup
 # gradle ignore
 .gradle
 build/
+out/
 
 # Ignore Gradle GUI config
 gradle-app.setting
